+++ common/src/main/java/net/mehvahdjukaar/jeed/api/JeedAPI.java
@@ -1,11 +1,19 @@
 package net.mehvahdjukaar.jeed.api;
 
+import net.mehvahdjukaar.jeed.Jeed;
+import net.mehvahdjukaar.jeed.common.EffectRenderer;
 import net.mehvahdjukaar.jeed.common.ScreenExtensionsHandler;
-import net.minecraft.client.gui.screens.Screen;
 import net.minecraft.client.gui.screens.inventory.AbstractContainerScreen;
 import net.minecraft.client.gui.screens.inventory.CreativeModeInventoryScreen;
 import net.minecraft.client.gui.screens.inventory.EffectRenderingInventoryScreen;
 import net.minecraft.client.gui.screens.inventory.InventoryScreen;
+import net.minecraft.core.Holder;
+import net.minecraft.network.chat.Component;
+import net.minecraft.world.effect.MobEffect;
+import net.minecraft.world.effect.MobEffectInstance;
+import net.minecraft.world.item.TooltipFlag;
+
+import java.util.List;
 
 /**
  * Register here extensions for your Screen classes where effects are rendered.
@@ -19,7 +27,7 @@ public class JeedAPI {
      * You can pass here a superclasses, and it will apply to all its subclasses
      **/
     public static <T extends AbstractContainerScreen<?>> void registerScreenExtension(Class<T> screenClass, IEffectScreenExtension<T> extension) {
-        ScreenExtensionsHandler.registerScreenExtension(screenClass,extension);
+        ScreenExtensionsHandler.registerScreenExtension(screenClass, extension);
     }
 
     /**
@@ -36,6 +44,21 @@ public class JeedAPI {
     public static <T extends AbstractContainerScreen> void disableExtension(Class<T> screenClass) {
         ScreenExtensionsHandler.unRegisterExtension(screenClass);
     }
+
+    // Delegate calls. Just here for clarity
+
+    public static void invokeEffectClicked(MobEffectInstance effectInstance, double mouseX, double mouseY, int button) {
+        Jeed.PLUGIN.onClickedEffect(effectInstance, mouseX, mouseY, button);
+    }
+
+    public static List<Component> getEffectTooltip(MobEffectInstance effectInstance, TooltipFlag tooltipFlag,
+                                                   boolean reactsToShift, boolean showDuration) {
+        return EffectRenderer.getTooltipsWithDescription(effectInstance, tooltipFlag, reactsToShift, showDuration);
+    }
+
+    public static boolean isHiddenEffect(Holder<MobEffect> effectInstance) {
+        return effectInstance.is(Jeed.HIDDEN);
+}
 
 
 }