+++ common/src/main/java/net/mehvahdjukaar/jeed/recipes/EffectProviderRecipe.java
@@ -7,9 +7,7 @@ import com.mojang.serialization.MapCodec;
 import com.mojang.serialization.codecs.RecordCodecBuilder;
 import net.mehvahdjukaar.jeed.Jeed;
 import net.mehvahdjukaar.jeed.common.CodecUtil;
-import net.minecraft.core.Holder;
-import net.minecraft.core.HolderLookup;
-import net.minecraft.core.NonNullList;
+import net.minecraft.core.*;
 import net.minecraft.core.component.DataComponentPatch;
 import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.core.registries.Registries;
@@ -21,6 +19,7 @@ import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.crafting.*;
 import net.minecraft.world.item.trading.MerchantOffer;
 import net.minecraft.world.level.Level;
+import net.minecraft.world.level.material.Fluid;
 
 import java.util.Collection;
 import java.util.Collections;
@@ -29,7 +28,8 @@ import java.util.Optional;
 
 public record EffectProviderRecipe(Optional<Holder<MobEffect>> effect,
                                    NonNullList<Ingredient> providers,
-                                   List<Holder<MobEffect>> effectProviders)
+                                   HolderSet<MobEffect> effectProviders,
+                                   HolderSet<Fluid> fluidProviders)
         implements Recipe<SingleRecipeInput> {
 
     @Override
@@ -92,13 +92,15 @@ public record EffectProviderRecipe(Optional<Holder<MobEffect>> effect,
                     Ingredient[] ingredients = list.stream().filter((ingredient) -> !ingredient.isEmpty()).toArray(Ingredient[]::new);
                     return DataResult.success(NonNullList.of(Ingredient.EMPTY, ingredients));
                 }, DataResult::success).forGetter((shapelessRecipe) -> shapelessRecipe.providers),
-                BuiltInRegistries.MOB_EFFECT.holderByNameCodec().listOf().optionalFieldOf("effect_providers", List.of()).forGetter((r) -> r.effectProviders)
+                RegistryCodecs.homogeneousList(Registries.MOB_EFFECT).optionalFieldOf("effect_providers", HolderSet.empty()).forGetter((r) -> r.effectProviders),
+                RegistryCodecs.homogeneousList(Registries.FLUID).optionalFieldOf("fluid_providers", HolderSet.empty()).forGetter((r) -> r.fluidProviders)
         ).apply(instance, EffectProviderRecipe::new));
 
         public static final StreamCodec<RegistryFriendlyByteBuf, EffectProviderRecipe> STREAM_CODEC = StreamCodec.composite(
                 ByteBufCodecs.optional(ByteBufCodecs.holderRegistry(Registries.MOB_EFFECT)), EffectProviderRecipe::effect,
                 Ingredient.CONTENTS_STREAM_CODEC.apply(ByteBufCodecs.collection(NonNullList::createWithCapacity)), EffectProviderRecipe::providers,
-                ByteBufCodecs.holderRegistry(Registries.MOB_EFFECT).apply(ByteBufCodecs.list()), EffectProviderRecipe::effectProviders,
+                ByteBufCodecs.holderSet(Registries.MOB_EFFECT), EffectProviderRecipe::effectProviders,
+                ByteBufCodecs.holderSet(Registries.FLUID), EffectProviderRecipe::fluidProviders,
                 EffectProviderRecipe::new);
 
         @Override